re-organize api
